# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PROSIC – Plataforma de Logística Inteligente
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Sesión iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Credenciales inválidas
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Obtener usuario actual
      responses:
        "200":
          description: Usuario autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: No autenticado

  # ─── CLIENTES ────────────────────────────────────────────────────────────
  /clients:
    get:
      operationId: listClients
      tags: [catalogs]
      summary: Listar clientes
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sector
          schema:
            type: string
      responses:
        "200":
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [catalogs]
      summary: Crear cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [catalogs]
      summary: Obtener cliente
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    put:
      operationId: updateClient
      tags: [catalogs]
      summary: Actualizar cliente
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [catalogs]
      summary: Eliminar cliente
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cliente eliminado

  # ─── VEHÍCULOS ───────────────────────────────────────────────────────────
...
  /equipment:
    get:
      operationId: listEquipment
      tags: [catalogs]
      summary: Listar equipos
      parameters:
        - in: query
          name: tipo
          schema:
            type: string
      responses:
        "200":
          description: Lista de equipos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: createEquipment
      tags: [catalogs]
      summary: Crear equipo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Equipo creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    put:
      operationId: updateEquipment
      tags: [catalogs]
      summary: Actualizar equipo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpdate"
      responses:
        "200":
          description: Equipo actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [catalogs]
      summary: Eliminar equipo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado

  # ─── CONDUCTORES ─────────────────────────────────────────────────────────
  /drivers:
    get:
      operationId: listDrivers
      tags: [catalogs]
      summary: Listar conductores / proveedores
      parameters:
        - in: query
          name: tipo
          schema:
            type: string
            enum: [PROPIO, EXTERNO]
      responses:
        "200":
          description: Lista de conductores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [catalogs]
      summary: Crear conductor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Conductor creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    put:
      operationId: updateDriver
      tags: [catalogs]
      summary: Actualizar conductor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdate"
      responses:
        "200":
          description: Conductor actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
    delete:
      operationId: deleteDriver
      tags: [catalogs]
      summary: Eliminar conductor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado

  # ─── SUCURSALES ───────────────────────────────────────────────────────────
...
  /quotes:
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Generar cotización automática
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequest"
      responses:
        "200":
          description: Cotización generada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"

    get:
      operationId: listQuotes
      tags: [quotes]
      summary: Listar cotizaciones
      parameters:
        - in: query
          name: estatus
          schema:
            type: string
            enum: [BORRADOR, ENVIADA, ACEPTADA, RECHAZADA, EXPIRADA]
      responses:
        "200":
          description: Cotizaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Obtener cotización
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cotización
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/postal-zones:
    get:
      operationId: listPostalZones
      tags: [quotes]
      summary: Zonas tarifarias por código postal
      parameters:
        - in: query
          name: cp
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Zona y tarifa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostalZone"
  /quotes/validate-identity:
    post:
      operationId: validateIdentity
      tags: [quotes]
      summary: Validar identidad persona física/moral
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentityValidationRequest"
      responses:
        "200":
          description: Resultado de validación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityValidationResult"

  # ─── OPTIMIZACIÓN DE RUTAS ────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthSession:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, nombre, rol]
      properties:
        id:
          type: integer
        email:
          type: string
        nombre:
          type: string
        rol:
          type: string
          enum: [OPERACIONES, TRAFICO, VENDEDOR, GERENCIA, ADMIN]
        sucursal:
          type: ["string", "null"]

    # CLIENTES
    Client:
      type: object
      required: [id, cliente, obra]
      properties:
        id:
          type: integer
        cliente:
          type: string
        obra:
          type: string
        comentarios_operaciones:
          type: ["string", "null"]
        calificacion:
          type: ["string", "null"]
          enum: [bueno, regular, malo, null]
        sector_alimenticio:
          type: boolean
        sector_automotriz:
          type: boolean
        sector_construccion:
          type: boolean
        created_at:
          type: string
          format: date-time
    ClientInput:
      type: object
      required: [cliente, obra]
      properties:
        cliente:
          type: string
        obra:
          type: string
        comentarios_operaciones:
          type: string
        calificacion:
          type: string
          enum: [bueno, regular, malo]
        sector_alimenticio:
          type: boolean
        sector_automotriz:
          type: boolean
        sector_construccion:
          type: boolean
    ClientUpdate:
      type: object
      properties:
        cliente:
          type: string
        obra:
          type: string
        comentarios_operaciones:
          type: string
        calificacion:
          type: string
        sector_alimenticio:
          type: boolean
        sector_automotriz:
          type: boolean
        sector_construccion:
          type: boolean

    # VEHÍCULOS
...
    Equipment:
      type: object
      required: [id, modelo, serie]
      properties:
        id:
          type: integer
        id_equipo:
          type: string
        modelo:
          type: string
        serie:
          type: string
        tipo:
          type: ["string", "null"]
        disponible:
          type: boolean
    EquipmentInput:
      type: object
      required: [modelo, serie]
      properties:
        modelo:
          type: string
        serie:
          type: string
        tipo:
          type: string
    EquipmentUpdate:
      type: object
      properties:
        modelo:
          type: string
        serie:
          type: string
        tipo:
          type: string
        disponible:
          type: boolean

    # CONDUCTORES
    Driver:
      type: object
      required: [id, nombre, tipo]
      properties:
        id:
          type: integer
        id_conductor:
          type: string
        nombre:
          type: string
        tipo:
          type: string
          enum: [PROPIO, EXTERNO]
        activo:
          type: boolean
        servicios_mes:
          type: integer
    DriverInput:
      type: object
      required: [nombre, tipo]
      properties:
        nombre:
          type: string
        tipo:
          type: string
          enum: [PROPIO, EXTERNO]
    DriverUpdate:
      type: object
      properties:
        nombre:
          type: string
        tipo:
          type: string
          enum: [PROPIO, EXTERNO]
        activo:
          type: boolean

    # SUCURSALES
...
    QuoteRequest:
      type: object
      required: [codigo_postal, modelo, dias_renta, tipo_persona]
      properties:
        codigo_postal:
          type: string
        modelo:
          type: string
        serie:
          type: string
        dias_renta:
          type: integer
        tipo_persona:
          type: string
          enum: [FISICA, MORAL]
        nombre_empresa:
          type: string
        rfc:
          type: string
        nombre_contacto:
          type: string
        telefono:
          type: string
        email:
          type: string
        direccion_entrega:
          type: string
        notas:
          type: string
    Quote:
      type: object
      required: [id, estatus, monto_total]
      properties:
        id:
          type: integer
        folio:
          type: string
        codigo_postal:
          type: string
        zona:
          type: string
        modelo:
          type: string
        serie:
          type: ["string", "null"]
        dias_renta:
          type: integer
        tipo_persona:
          type: string
          enum: [FISICA, MORAL]
        nombre_empresa:
          type: ["string", "null"]
        rfc:
          type: ["string", "null"]
        nombre_contacto:
          type: ["string", "null"]
        telefono:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        costo_flete:
          type: number
        costo_renta_diaria:
          type: number
        subtotal:
          type: number
        iva:
          type: number
        monto_total:
          type: number
        estatus:
          type: string
          enum: [BORRADOR, ENVIADA, ACEPTADA, RECHAZADA, EXPIRADA]
        identidad_verificada:
          type: boolean
        validacion_resultado:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        vigencia:
          type: string
          format: date-time
    PostalZone:
      type: object
      required: [codigo_postal, zona, tarifa_flete]
      properties:
        codigo_postal:
          type: string
        municipio:
          type: string
        estado:
          type: string
        zona:
          type: string
        tarifa_flete:
          type: number
        tiempo_estimado_hrs:
          type: number
    IdentityValidationRequest:
      type: object
      required: [tipo_persona, rfc]
      properties:
        tipo_persona:
          type: string
          enum: [FISICA, MORAL]
        rfc:
          type: string
        nombre:
          type: string
        curp:
          type: string
    IdentityValidationResult:
      type: object
      required: [valido, tipo_persona]
      properties:
        valido:
          type: boolean
        tipo_persona:
          type: string
          enum: [FISICA, MORAL]
        nombre_registrado:
          type: ["string", "null"]
        rfc_verificado:
          type: boolean
        observaciones:
          type: ["string", "null"]

    # RUTAS
...